feat(add try catch to controller): add try catch to controller
Add try catch to controller, refactor validators, change button data-cy attr
