+++ packages/forms-web-app/src/validators/shared/index.js
@@ -1,15 +1,5 @@
-const { body } = require('express-validator');
-
-const validateNotEmptyAndLength = ({ notEmpty, checkLength, id, options }) => {
-	return [
-		body(id).notEmpty().withMessage(notEmpty),
-		body(id).isLength(options).withMessage(checkLength)
-	];
-};
-
-const validateNotEmpty = (id, message) => {
-	return [body(id).notEmpty().withMessage(message)];
-};
+const { validateNotEmpty } = require('./not-empty');
+const { validateNotEmptyAndLength } = require('./not-empty-and-length');
 
 module.exports = {
 	validateNotEmpty,
