+++ packages/forms-web-app/src/validators/shared/not-empty.js
@@ -1,0 +1,9 @@
+const { body } = require('express-validator');
+
+const validateNotEmpty = (id, message) => {
+	return [body(id).notEmpty().withMessage(message)];
+};
+
+module.exports = {
+	validateNotEmpty
+};
