+++ packages/forms-web-app/src/views/pages/examination/your-interested-party-number.njk
@@ -44,7 +44,7 @@
 			{{ govukButton({
 				text: "Continue",
 				type: "Submit",
-				attributes: { "data-cy":"button-save-and-continue"}
+				attributes: { "data-cy":"button-submit-and-continue"}
 			}) }}
 
 		</form>
